brute-force-plist/requirements.txt
fastpbkdf2
pycrypto
